Backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
motor>=3.3.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
email-validator>=2.1.0
fastapi-mail>=1.4.1
httpx>=0.25.0

# ISO 9001 Analysis System
langchain==0.1.0
langchain-community==0.0.13
langchain-groq==0.0.1
chromadb==0.4.22
sentence-transformers==2.3.1
pypdf==4.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
tiktoken==0.5.2
numpy==1.26.3
pandas==2.1.4
tqdm==4.66.1

# Document extraction
pypdf2>=3.0.0
python-docx>=1.1.0

# Text processing
pydantic>=2.5.0
